requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via
    #   jsonschema
    #   openedx-events
    #   referencing
avro==1.11.2
    # via confluent-kafka
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   code-annotations
    #   edx-django-utils
code-annotations==1.5.0
    # via edx-toggles
confluent-kafka[avro,schema-registry]==2.2.0
    # via -r requirements/base.in
cryptography==41.0.3
    # via
    #   -r requirements/base.in
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.20
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-model-utils
    #   django-notifications-hq
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   drf-spectacular
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-event-bus-kafka
    #   edx-toggles
    #   jsonfield
    #   openedx-events
    #   social-auth-app-django
django-cors-headers==4.2.0
    # via -r requirements/base.in
django-crum==0.7.9
    # via
    #   edx-django-utils
    #   edx-toggles
django-extensions==3.2.3
    # via -r requirements/base.in
django-model-utils==4.3.1
    # via django-notifications-hq
django-notifications-hq==1.8.2
    # via -r requirements/base.in
django-simple-history==3.0.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
django-waffle==4.0.0
    # via
    #   -r requirements/base.in
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   drf-jwt
    #   drf-spectacular
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via edx-drf-extensions
drf-spectacular==0.26.4
    # via -r requirements/base.in
edx-auth-backends==4.1.0
    # via -r requirements/base.in
edx-django-release-util==1.3.0
    # via -r requirements/base.in
edx-django-utils==5.6.0
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-event-bus-kafka
    #   edx-rest-api-client
    #   edx-toggles
edx-drf-extensions==8.8.0
    # via -r requirements/base.in
edx-event-bus-kafka==5.1.0
    # via -r requirements/base.in
edx-opaque-keys[django]==2.3.0
    # via
    #   edx-drf-extensions
    #   openedx-events
edx-rest-api-client==5.6.0
    # via -r requirements/base.in
edx-toggles==5.1.0
    # via edx-event-bus-kafka
fastavro==1.8.2
    # via
    #   confluent-kafka
    #   openedx-events
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jinja2==3.1.2
    # via code-annotations
jsonfield==3.1.0
    # via django-notifications-hq
jsonschema==4.18.5
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
markupsafe==2.1.3
    # via jinja2
mysqlclient==2.2.0
    # via -r requirements/base.in
newrelic==8.9.0
    # via edx-django-utils
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
openedx-events==8.4.0
    # via
    #   -r requirements/base.in
    #   edx-event-bus-kafka
pbr==5.11.1
    # via stevedore
pillow==10.0.0
    # via -r requirements/base.in
psutil==5.9.5
    # via edx-django-utils
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pymongo==3.13.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
python-dateutil==2.8.2
    # via edx-drf-extensions
python-slugify==8.0.1
    # via code-annotations
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3
    # via
    #   -r requirements/base.in
    #   django
    #   django-notifications-hq
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   code-annotations
    #   drf-spectacular
    #   edx-django-release-util
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   confluent-kafka
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
semantic-version==2.10.0
    # via edx-drf-extensions
six==1.16.0
    # via
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   python-dateutil
slumber==0.7.1
    # via edx-rest-api-client
social-auth-app-django==5.2.0
    # via edx-auth-backends
social-auth-core==4.4.2
    # via
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.4
    # via django
stevedore==5.1.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
swapper==1.3.0
    # via django-notifications-hq
text-unidecode==1.3
    # via python-slugify
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.0.4
    # via requests
